Remove remote write code from REST API files
