+++ exporting/prometheus/remote_write/remote_write.c
@@ -261,3 +261,24 @@ int format_batch_prometheus_remote_write(struct instance *instance)
 
     return 0;
 }
+
+int process_prometheus_remote_write_response(BUFFER *b, struct instance *instance) {
+    if(unlikely(!b)) return 1;
+
+    const char *s = buffer_tostring(b);
+    int len = buffer_strlen(b);
+
+    // do nothing with HTTP responses 200 or 204
+
+    while(!isspace(*s) && len) {
+        s++;
+        len--;
+    }
+    s++;
+    len--;
+
+    if(likely(len > 4 && (!strncmp(s, "200 ", 4) || !strncmp(s, "204 ", 4))))
+        return 0;
+    else
+        return exporting_discard_response(b, instance);
+}
